+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.2.1 (2019-07-18)
 
 - FEAT: drop deprecated piece of code in pull.py that was moved into cli_*.py
 - FEAT: use jmespath in filtering the cloudwatch describe alarms response
+- FEAT: implement cloudwatch list-metrics remote fetch
 
 
 Version 0.2.0 (2019-07-16)